Consolidate styles
