--- components/FunctionNavigation.styles.js
@@ -1,17 +1,0 @@
-import css from 'styled-jsx/css';
-
-const styles = css`
-  a.function-link {
-    display: inline-block;
-    width: 100%;
-    padding: 0.8rem 0;
-    text-decoration: none;
-    color: #333;
-  }
-
-  a.function-link:hover {
-    color: #999;
-  }
-`;
-
-export default styles;
