+++ md/FunctionsOverview.mdx
@@ -1,5 +1,3 @@
-import FunctionList from '../components/FunctionList';
-
 # Functions
 
 Functions describe responsibilities that engineers may be expected to fulfill for a period of time.
